first working version of PIOP1 and PIOP2 plus tests and testing datasets on gin
